コードの切り出し
